moves all per tab to new component
